+++ scikit_quri/state/overlap_estimator.py
@@ -1,81 +1,47 @@
-from typing import List, Sequence
-from quri_parts.core.state import GeneralCircuitQuantumState, QuantumState
-import numpy as np
-from numpy.typing import NDArray
-from quri_parts.qulacs.circuit import convert_circuit
-from quri_parts.qulacs.overlap_estimator import (
-    create_qulacs_vector_overlap_estimator,
-    _create_qulacs_initial_state,
-)
-from qulacs.state import inner_product
-from qulacs import QuantumState as QulacsQuantumState
-from quri_parts.core.state import GeneralCircuitQuantumState
+from typing import List
+from quri_parts.circuit import QuantumCircuit
+from quri_parts.circuit.inverse import inverse_circuit, inverse_gate
+from quri_parts.backend import SamplingBackend
+from quri_parts.core.sampling import Sampler
 
 
+# TODO cache化
 class overlap_estimator:
     """
     quri-partsのoverlap_estimatorの代替Class
     (n_data:500のとき,x60 faster)
     """
 
-    def __init__(self, states: List[QuantumState]):
+    def __init__(self, sampler: Sampler, n_shots: int = 1000):
         """
         Args:
-            states (List[GeneralCircuitQuantumState]): 量子状態のリスト
+            circuits (List[QuantumCircuit]): 量子回路のリスト
         """
-        self.states = states
-        self.qula_states: NDArray[np.float64] = np.full(len(states), fill_value=None, dtype=object)
+        self.sampler = sampler
+        self.n_shots = n_shots
+        self.cache = {}
 
-    def _state_to_qula_state(self, state: GeneralCircuitQuantumState) -> QulacsQuantumState:
-        """
-        量子状態をqulacsのstateに変換
-        Args:
-            state (GeneralCircuitQuantumState): quri-partsの量子状態
-        Returns:
-            qulacs_state (QuantumState): qulacsの量子状態
-        """
-        circuit = convert_circuit(state.circuit)
-        qulacs_state = _create_qulacs_initial_state(state)
-        circuit.update_quantum_state(qulacs_state)
-        return qulacs_state
-
-    def add_data(self, states: List[GeneralCircuitQuantumState]) -> None:
-        """
-        量子状態を追加
-        Args:
-            states (List[GeneralCircuitQuantumState]): 量子状態のリスト
-        """
-        self.states.extend(states)
-        self.qula_states = np.append(
-            self.qula_states, np.full(len(states), fill_value=None, dtype=object)
-        )
-
-    def calc_all_qula_states(self) -> None:
-        """
-        cache用に予め全ての量子状態をqulacsのstateに変換
-        """
-        for i in range(len(self.states)):
-            self.qula_states[i] = self._state_to_qula_state(self.states[i])
-
-    def estimate(self, i: int, j: int) -> float:
+    def estimate(self, ket_circuit: QuantumCircuit, bra_circuit: QuantumCircuit) -> float:
         # ? これi,jじゃなくて数値でhash取った方が使いやすそう
         """
-        与えられた量子状態のi番目とj番目の内積の絶対値の二乗を計算
+        与えられた量子回路のi番目とj番目の内積の絶対値の二乗を計算
+        引数に対して非破壊的に動作
         Args:
             i (int): 量子状態のindex(ket)
             j (int): 量子状態のindex(bra)
         Returns:
             float: |<φi|φj>|^2
         """
-        ket = self.qula_states[i]
-        # qulacsのstateを使いまわす
-        if ket is None:
-            ket = self._state_to_qula_state(self.states[i])
-            self.qula_states[i] = ket
-        bra = self.qula_states[j]
-        if bra is None:
-            bra = self._state_to_qula_state(self.states[j])
-            self.qula_states[j] = bra
-        overlap = inner_product(bra, ket)
-        overlap_mag_sqrd = abs(overlap) ** 2
-        return overlap_mag_sqrd
+        ket_circuit = ket_circuit.get_mutable_copy()
+        bra_circuit = bra_circuit.get_mutable_copy()
+        circuit = ket_circuit
+        # 逆順にcombine
+        inv_circuit = inverse_circuit(bra_circuit)
+        for inv_gate in inv_circuit.gates:
+            circuit.add_gate(inv_gate)
+        sampling_count = self.sampler(circuit, self.n_shots)
+        count_zero = sampling_count.get(0)
+        if count_zero is None:
+            count_zero = 0
+        p = count_zero / self.n_shots
+        return p
